split TravelTimeMatrixComputer into a parent-child class architecture
